bot/core/requirements.txt
discord.py
yt-dlp[default]
yt-dlp-ejs
PyNaCl
openai
mutagen
requests
python-dotenv
pytz
cycler
dotenv
ffmpeg
py7zr
gTTS
httpcore
matplotlib
numpy
pylance
yarl
aiohttp